+++ packages/expo-router/CHANGELOG.md
@@ -44,6 +44,7 @@
 - [iOS] add Toolbar.SearchBarPreferredSlot ([#41863](https://github.com/expo/expo/pull/41863) by [@Ubax](https://github.com/Ubax))
 - [iOS] add usePreventZoomTransitionDismissal ([#41893](https://github.com/expo/expo/pull/41893) by [@Ubax](https://github.com/Ubax))
 - [iOS] expose image prop from Link.Menu and Toolbar.Menu ([#41883](https://github.com/expo/expo/pull/41883) by [@Ubax](https://github.com/Ubax))
+- add platform theme utility ([#40905](https://github.com/expo/expo/pull/40905) by [@Ubax](https://github.com/Ubax))
 
 ### 🐛 Bug fixes